requirements.txt
Pillow
mysql-connector
fpdf